Bug fix: zIndex of the rectangle border of the board.
The zIndex of the border was not set (default = 0). So, the chip was being drawn on top of it.
Made it equal to 19 (the zIndex of the board itself.)

modified:   src/main/java/connectXviewer/Viewer.java
modified:   src/main/resources/view/demo.js

Updated intro replay so that it is not shown in previous way.