Bug fix: Stolen chip is shown smaller sometimes
modified:   src/main/java/connectXviewer/Viewer.java

How to reproduce bug in previous version:
Use mouse to drag in the viewer's timeline from the beginning
when the steal animation is in midway, stop the mouse drag
Now click on "previous frame" (<), or, "first frame" (<<) buttons under the timeline
Now go forawd in timeline by clicking on "next frame" (>) button
The Stolen chip is shown as same size as when the mouse drag stopped.

How it is fixed:
A new full chip is drawn and committed at the end of the frame.